LPS-95267 Move logic to renamed tag
