+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/start.jsp
@@ -14,4 +14,17 @@
  */
 --%>
 
-<%@ include file="init.jsp" %>
+<%@ include file="/data_layout_builder/init.jsp" %>
+
+<liferay-util:html-top>
+	<link href="<%= PortalUtil.getStaticResourceURL(request, "/o/dynamic-data-mapping-form-builder/css/main.css") %>" rel="stylesheet" type="text/css" />
+</liferay-util:html-top>
+
+<div class="container-fluid-1280 ddm-translation-manager">
+	<liferay-frontend:translation-manager
+		availableLocales="<%= availableLocales %>"
+		changeableDefaultLanguage="<%= false %>"
+		defaultLanguageId="<%= themeDisplay.getLanguageId() %>"
+		id="translationManager"
+	/>
+</div>
