--- modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/layout_builder/init-ext.jspf
@@ -1,40 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<%@ page import="com.liferay.portal.kernel.json.JSONArray" %><%@
-page import="com.liferay.portal.kernel.json.JSONObject" %><%@
-page import="com.liferay.portal.kernel.util.GetterUtil" %><%@
-page import="com.liferay.portal.kernel.util.PortalUtil" %>
-
-<%@ page import="java.util.Locale" %><%@
-page import="java.util.Map" %>
-
-<%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
-taglib uri="http://liferay.com/tld/frontend" prefix="liferay-frontend" %><%@
-taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
-taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
-
-<liferay-frontend:defineObjects />
-
-<liferay-theme:defineObjects />
-
-<%
-	Locale[] availableLocales = (Locale[])request.getAttribute("liferay-data-engine:layout-builder:availableLocales");
-	JSONArray fieldTypes = (JSONArray)request.getAttribute("liferay-data-engine:layout-builder:fieldTypes");
-	String fieldTypesModules = (String)request.getAttribute("liferay-data-engine:layout-builder:fieldTypesModules");
-	JSONObject dataLayout = (JSONObject)request.getAttribute("liferay-data-engine:layout-builder:dataLayout");
-	String layoutBuilderModule = (String)request.getAttribute("liferay-data-engine:layout-builder:layoutBuilderModule");
-%>
